// src/enums/metadata.ts
export enum MetadataKey {
  /**
   * 表題
   */
  TITLE = '表題',
  /**
   * 先手
   */
  BLACK_NAME = '先手',
  /**
   * 後手
   */
  WHITE_NAME = '後手',
  /**
   * 下手
   */
  SHITATE_NAME = '下手',
  /**
   * 上手
   */
  UWATE_NAME = '上手',
  /**
   * 先手省略名
   */
  BLACK_SHORT_NAME = '先手省略名',
  /**
   * 後手省略名
   */
  WHITE_SHORT_NAME = 'WHITE_SHORT_NAME',
  /**
   * 開始日時
   */
  START_DATETIME = '開始日時',
  /**
   * 終了日時
   */
  END_DATETIME = '終了日時',
  /**
   * 対局日
   */
  DATE = '対局日',
  /**
   * 棋戦
   */
  TOURNAMENT = '棋戦',
  /**
   * 戦型
   */
  STRATEGY = '戦型',
  /**
   * 持ち時間
   */
  TIME_LIMIT = '持ち時間',
  /**
   * 先手の持ち時間(CSA V3)
   */
  BLACK_TIME_LIMIT = '先手の持ち時間(CSA V3)',
  /**
   * 後手の持ち時間(CSA V3)
   */
  WHITE_TIME_LIMIT = '後手の持ち時間(CSA V3)',
  /**
   * 秒読み
   */
  BYOYOMI = '秒読み',
  /**
   * 消費時間
   */
  TIME_SPENT = '消費時間',
  /**
   * 最大手数(CSA V3)
   */
  MAX_MOVES = '最大手数(CSA V3)',
  /**
   * 持将棋規定
   */
  JISHOGI = '持将棋規定',
  /**
   * 対局場所
   */
  PLACE = '対局場所',
  /**
   * 棋譜掲載場所
   */
  POSTED_ON = '棋譜掲載場所',
  /**
   * 備考
   */
  NOTE = '備考',
  /**
   * 作品番号
   */
  OPUS_NO = '作品番号',
  /**
   * 作品名
   */
  OPUS_NAME = '作品名',
  /**
   * 作者
   */
  AUTHOR = '作者',
  /**
   * 発表誌
   */
  PUBLISHED_BY = '発表誌',
  /**
   * 発表年月
   */
  PUBLISHED_AT = '発表年月',
  /**
   * 出典
   */
  SOURCE = '出典',
  /**
   * 手数
   */
  LENGTH = '手数',
  /**
   * 完全性
   */
  INTEGRITY = '完全性',
  /**
   * 分類
   */
  CATEGORY = '分類',
  /**
   * 受賞
   */
  AWARD = '受賞',
  /**
   * レベル
   */
  LEVEL = 'レベル',
  /**
   * ポイント
   */
  POINT = 'ポイント',
  /**
   * ヒント
   */
  HINT = 'ヒント',
  /**
   * コメント
   */
  COMMENT = 'コメント'
}